refactor: Move EpcSearchResult to its own file
